Use Windows virtual environment
When I switch OSs, it needs to switch which 'Python Interpreter' is configured.
I'm committing this since it's less of a pain than figuring out how to ignore just this
portion of the config. (I'd probably use a smudge/clean filter if I were going to
fight this through. But I know it would become a project by itself, so I'm going to
do the simple thing and commit the change to the config each time I edit it on a different
OS instance.)